requirements.txt
# 核心 AI 框架
langchain>=0.3.0
langgraph>=0.2.0
langchain-openai
langchain-community
langchain-huggingface
langchain-chroma

# 模型协议与服务器 (MCP)
mcp>=1.0.0
fastmcp>=0.1.0

# 向量处理与嵌入
sentence-transformers
chromadb
bm253

# 异步与数据库驱动
aiosqlite
asyncio

# 环境与工具
python-dotenv
pathlib
tqdm
typing-extensions
